requirements.txt
psycopg2==2.8.4
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1